oppdaterte package script
